refactor: 移動によるAI関連コンポーネントと型のパス整理
- AIInsights, AIControlPanel, PerformanceAnalysis, StrategySettingsコンポーネントを`ai`ディレクトリに移動
- 関連する型定義やサービスのパスを調整
- import文を新しい構成にあわせて更新して整理
- 組織化を改善しコードの可読性を向上
