+++ app/src/App.tsx
@@ -6,14 +6,14 @@ import type { InputPort } from './application/ports/InputPort'
 import type { PerformanceAnalysisService } from './application/services/PerformanceAnalysisService'
 import type { StrategyService } from './application/services/ai/StrategyService'
 import type { GameViewModel } from './application/viewmodels/GameViewModel'
-import type { AIMove, AISettings } from './domain/models/ai/index'
+import type { AIMove, AISettings } from './domain/models/ai'
 import { defaultContainer } from './infrastructure/di/DefaultContainer'
-import { AIControlPanel } from './presentation/components/AIControlPanel'
-import { AIInsights } from './presentation/components/AIInsights'
+import { AIControlPanel } from './presentation/components/ai/AIControlPanel'
+import { AIInsights } from './presentation/components/ai/AIInsights'
 import { GameBoard } from './presentation/components/GameBoard'
 import { GameInfo } from './presentation/components/GameInfo'
-import { PerformanceAnalysis } from './presentation/components/PerformanceAnalysis'
-import { StrategySettings } from './presentation/components/StrategySettings'
+import { PerformanceAnalysis } from './presentation/components/ai/PerformanceAnalysis'
+import { StrategySettings } from './presentation/components/ai/StrategySettings'
 import { useAutoFall } from './presentation/hooks/useAutoFall'
 import { useKeyboard } from './presentation/hooks/useKeyboard'
 import { usePerformanceAnalysis } from './presentation/hooks/usePerformanceAnalysis'
