+++ app/src/presentation/components/ai/PerformanceAnalysis.test.tsx
@@ -5,8 +5,8 @@ import { describe, expect, test, vi } from 'vitest'
 
 import { render, screen } from '@testing-library/react'
 
-import type { PerformanceStatistics } from '../../application/services/PerformanceAnalysisService'
-import type { PerformanceReport } from '../../domain/models/ai/index'
+import type { PerformanceStatistics } from '../../../application/services/PerformanceAnalysisService'
+import type { PerformanceReport } from '../../../domain/models/ai'
 import { PerformanceAnalysis } from './PerformanceAnalysis'
 
 describe('PerformanceAnalysis', () => {
