+++ app/src/presentation/components/ai/PerformanceAnalysis.tsx
@@ -3,11 +3,11 @@
  */
 import { useState } from 'react'
 
-import type { PerformanceStatistics } from '../../application/services/PerformanceAnalysisService'
-import { ChartDataService } from '../../application/services/visualization/ChartDataService'
-import type { PerformanceMetricType } from '../../application/services/visualization/ChartDataService'
-import type { PerformanceReport } from '../../domain/models/ai/index'
-import { LineChart } from './charts/LineChart'
+import type { PerformanceStatistics } from '../../../application/services/PerformanceAnalysisService'
+import { ChartDataService } from '../../../application/services/visualization/ChartDataService'
+import type { PerformanceMetricType } from '../../../application/services/visualization/ChartDataService'
+import type { PerformanceReport } from '../../../domain/models/ai'
+import { LineChart } from '../charts/LineChart'
 
 interface PerformanceAnalysisProps {
   statistics: PerformanceStatistics
